requirements.txt
fastapi
uvicorn
pyrogram
tgcrypto
python-dotenv
pydantic
pydantic-settings
httpx
aiosqlite
aiogram
pytz 
python-telegram-bot